Time: 27 ms (27.24%) | Memory: 92.6 MB (42.78%) - LeetSync

+++ 4011-smallest-absent-positive-greater-than-average/smallest-absent-positive-greater-than-average.cpp
@@ -10,8 +10,8 @@ public:
             st.insert(x);
         }
 
-        double avg = (double)sum / n;
-        int candidate = (int)floor(avg) + 1;   
+        int avg = sum / n;
+        int candidate = floor(avg) + 1;   
         candidate = max(candidate, 1);         
 
         while (st.count(candidate)) {
